I added game logic
I added comments that would help me understand how to check if the game is over. I checked the API which suggest "over" and if that is true or not. I am matching the player_x and player_o against the different methods of winning and if the array length is equal to 9 and nobody has won then it is a tie game.