I got my update curl script to work and am now working the rest of the event, the listener works and i changed the html format to match the game api.
lots of updates here, but the bottom line is that I successfully updated the game data with a curl script and set up an event listener that was successful. Now I am working on the ajax call and event.js/api.js to get everything ironed out. I also changed my html to reflect what was happening in the game api. Next up is incorporating game logic into the events.js file.