Updated ui, README, html
I made a few updates to make the ui make more sense. This functionality was pretty simple which include a few changes to classes and id's in html in order to allow me to better write my ui. I also wrote almost the entirety of my README and need to include a link with my wireframes and user stories.